+++ lobster/tools/core/report/report.py
@@ -18,13 +18,12 @@
 # <https://www.gnu.org/licenses/>.
 
 from argparse import Namespace
-import os
 import sys
 
 from lobster.exceptions import LOBSTER_Exception
 from lobster.errors import LOBSTER_Error
-from lobster.report import Report
 from lobster.meta_data_tool_base import MetaDataToolBase
+from lobster.report import Report
 
 
 class ReportTool(MetaDataToolBase):
@@ -46,30 +45,32 @@ class ReportTool(MetaDataToolBase):
         )
 
     def _run_impl(self, options: Namespace) -> int:
-        if not os.path.isfile(options.lobster_config):
-            print(f"error: cannot read config file '{options.lobster_config}'")
-            return 1
-
-        if os.path.exists(options.out) and not os.path.isfile(options.out):
-            print(f"error: cannot write to '{options.out}'")
-            print(f"error: '{options.out}' exists and is not a file")
-            return 1
 
         report = Report()
 
         try:
             report.parse_config(options.lobster_config)
-        except LOBSTER_Error:
-            print(f"{self.name}: aborting due to earlier errors.")
-            return 1
+        except FileNotFoundError as e:
+            print(f"{e}")
+        except TypeError as e:
+            print(f"{e}")
+        except LOBSTER_Error as e:
+            raise LOBSTER_Exception(
+                f"{self.name}: aborting due to earlier errors."
+            ) from e
         except LOBSTER_Exception as err:
-            print(f"{self.name}: aborting due to earlier errors.")
-            print(f"{self.name}: Additional data for debugging:")
             err.dump()
-            return 1
+            raise
 
         report.write_report(options.out)
         return 0
+
+
+def generate_report_file(lobster_config_file: str, output_file: str) -> dict:
+    # This is an API function to run the lobster report tool
+    report = Report()
+    report.parse_config(lobster_config_file)
+    report.write_report(output_file)
 
 
 def main() -> int: