+++ CHANGELOG.md
@@ -5,7 +5,11 @@
 
 ### 0.13.3-dev
 
-
+* Introduced API function:
+  - `generate_report_file`:
+    This is API function for the tool `lobster-report` which takes lobster config file as input
+    and generates the lobster report.
+    This is similar to running the tool `lobster-report`.
 
 ### 0.13.2